+++ docs/CHANGELOG.md
@@ -27,6 +27,7 @@ VoxConvert:
    - Added `--crop` parameter that reduces the volumes to their real voxel sizes
    - Added `--split` option to cut volumes into smaller pieces
    - Added `--export-layers` to export all the layers of a scene into single files
+   - Added `--dump` to dump the scene graph of the input file
 
 VoxEdit: